backend_drf/requirements.txt
django
djangorestframework
gunicorn
whitenoise
dj-database-url
psycopg2-binary
python-decouple
